[backend] refactor generateUpdateMessage
